+++ components/title-section/index.tsx
@@ -1,16 +1,10 @@
-import React from 'react';
-
-import { formatIntFr } from '../../utils/helpers/formatting';
-import { Tag } from '../tag';
-import IsActiveTag from '../is-active-tag';
-import { UnitLegaleDescription } from '../unite-legale-description';
-import { IUniteLegale } from '../../models';
-import SocialMedia from '../social-media';
-import { PrintNever } from '../print-visibility';
-import MultipleSirenAlert from '../alerts/multiple-siren';
-import AssociationAdressAlert from '../alerts/association-adress';
+import Tabs from './tabs';
+import LateralMenu from './lateral-menu';
+import Title from './title';
 
 export enum FICHE {
+  DOCUMENTS = 'mes documents',
+  SIMULATIONS = 'mes simulations',
   INFORMATION = 'informations générales',
   JUSTIFICATIFS = 'justificatifs',
   ANNONCES = 'annonces',
@@ -19,213 +13,52 @@ export enum FICHE {
   ACTES = 'actes & statuts',
   DIVERS = 'conventions collectives',
 }
-interface IProps {
-  ficheType?: FICHE;
-  uniteLegale: IUniteLegale;
-}
 
-export const Tabs: React.FC<{
-  ficheType: FICHE;
-  siren: string;
-  estDirigeant: boolean;
-}> = ({ ficheType, siren, estDirigeant = false }) => (
-  <PrintNever>
-    <div className={`title-tabs ${estDirigeant ? 'dirigeant' : ''}`}>
-      <a
-        className={`${ficheType === FICHE.INFORMATION && 'active'}`}
-        href={`/entreprise/${siren}`}
-      >
-        Informations générales
-      </a>
-      <a
-        className={`${ficheType === FICHE.JUSTIFICATIFS && 'active'}`}
-        href={`/justificatif/${siren}`}
-        rel="nofollow"
-      >
-        Justificatif d’immatriculation
-      </a>
-      <a
-        className={`${ficheType === FICHE.DIRIGEANTS && 'active'}`}
-        href={`/dirigeants/${siren}`}
-        rel="nofollow"
-      >
-        Dirigeants
-      </a>
-      <a
-        className={`${ficheType === FICHE.ANNONCES && 'active'}`}
-        href={`/annonces/${siren}`}
-        rel="nofollow"
-      >
-        Annonces
-      </a>
-      <a
-        className={`${ficheType === FICHE.DIVERS && 'active'}`}
-        href={`/divers/${siren}`}
-        rel="nofollow"
-      >
-        Conventions collectives
-      </a>
-    </div>
+export const menu = [
+  {
+    title: 'Mes documents',
+    type: FICHE.DOCUMENTS,
+    path: (siren: string) => `/comptes/documents/${siren}`,
+    private: true,
+  },
+  {
+    title: 'Mes simulations',
+    type: FICHE.SIMULATIONS,
+    path: (siren: string) => `/comptes/simulateurs/${siren}`,
+    private: true,
+  },
+  {
+    title: 'Informations générales',
+    type: FICHE.INFORMATION,
+    path: (siren: string) => `/entreprise/${siren}`,
+    private: false,
+  },
+  {
+    title: 'Justificatif d’immatriculation',
+    type: FICHE.JUSTIFICATIFS,
+    path: (siren: string) => `/justificatif/${siren}`,
+    private: false,
+  },
+  {
+    title: 'Dirigeants',
+    type: FICHE.DIRIGEANTS,
+    path: (siren: string) => `/dirigeants/${siren}`,
+    private: false,
+  },
+  {
+    title: 'Annonces',
+    type: FICHE.ANNONCES,
+    path: (siren: string) => `/annonces/${siren}`,
+    private: false,
+  },
+  {
+    title: 'Conventions collectives',
+    type: FICHE.DIVERS,
+    path: (siren: string) => `/divers/${siren}`,
+    private: false,
+  },
+];
 
-    <style jsx>{`
-      .title-tabs {
-        display: flex;
-        flex-grow: 1;
-        font-size: 0.9rem;
-        border-bottom: 2px solid #dfdff1;
-        margin-bottom: 40px;
-      }
-      .title-tabs.dirigeant > a {
-        color: #000091;
-        font-weight: bold;
-        border-top-left-radius: 3px;
-        border-top-right-radius: 3px;
-        margin: 0 4px;
-        padding: 10px 5px;
-        border: 2px solid #dfdff1;
-        background-color: #efeffb;
-        margin-bottom: -2px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        text-align: ${estDirigeant ? 'left' : 'center'};
-        box-shadow: 0 -8px 5px -5px #dfdff1 inset;
-      }
-
-      .title-tabs > a:hover {
-        background-color: #dfdff1;
-      }
-
-      .title-tabs > a.active {
-        background-color: #fff;
-        border-bottom: 0;
-        box-shadow: none;
-      }
-
-      .title-tabs.dirigeant {
-        flex-direction: column;
-        border-bottom: none;
-        margin: 0;
-      }
-      .title-tabs.dirigeant > a {
-        color: #000091;
-        font-weight: bold;
-        background-color: #fff;
-        border-radius: 3px;
-        border: none;
-        margin: 0;
-        padding-left: 10px;
-        margin: 2px 20px 2px 0;
-        text-align: left;
-        justify-content: start;
-        box-shadow: none;
-        border: 1px solid transparent;
-      }
-      .title-tabs.dirigeant > a:hover {
-        background-color: #efeffb;
-      }
-
-      .title-tabs.dirigeant > a.active {
-        background-color: #dfdff1;
-        box-shadow: none;
-      }
-
-      @media only screen and (min-width: 1px) and (max-width: 450px) {
-        .title-tabs {
-          flex-direction: column;
-          border-bottom: 0;
-        }
-        .title-tabs > a {
-          margin: 3px;
-        }
-        .title-tabs > a.active {
-          background-color: #fff;
-          border-bottom: 2px solid #dfdff1;
-          box-shadow: none;
-        }
-      }
-    `}</style>
-  </PrintNever>
-);
-
-const Title: React.FC<IProps> = ({ uniteLegale }) => (
-  <div className="header-section">
-    <div className="title">
-      <MultipleSirenAlert uniteLegale={uniteLegale} />
-      <AssociationAdressAlert uniteLegale={uniteLegale} />
-      <h1>
-        <a href={`/entreprise/${uniteLegale.siren}`}>
-          {uniteLegale.nomComplet}
-        </a>
-      </h1>
-      <div>
-        <span>unité légale ‣ {formatIntFr(uniteLegale.siren)}</span>
-        <span>
-          {!uniteLegale.estDiffusible ? (
-            <>
-              <Tag>Non-diffusible</Tag>
-              <IsActiveTag isActive={null} isUniteLegale={true} />
-            </>
-          ) : (
-            <IsActiveTag
-              isActive={uniteLegale.estActive}
-              isUniteLegale={true}
-            />
-          )}
-        </span>
-      </div>
-    </div>
-    <SocialMedia siren={uniteLegale.siren} />
-    {!uniteLegale.estDiffusible ? (
-      <p>Les informations concernant cette entité ne sont pas publiques.</p>
-    ) : (
-      <UnitLegaleDescription uniteLegale={uniteLegale} />
-    )}
-
-    <style jsx>{`
-      .header-section {
-        display: block;
-      }
-
-      .title {
-        margin: 40px 0 0;
-        display: flex;
-        align-items: start;
-        flex-direction: column;
-        justify-content: center;
-      }
-      .title h1 {
-        margin: 0;
-        margin-bottom: 7px;
-        line-height: 2.5rem;
-      }
-      .title h1 > a {
-        margin: 0;
-        padding: 0;
-      }
-      .title > div > span {
-        color: #666;
-      }
-      .title > div > span:first-of-type {
-        font-variant: small-caps;
-        font-size: 1.1rem;
-      }
-
-      @media only screen and (min-width: 1px) and (max-width: 900px) {
-        .title {
-          margin-top: 10px;
-        }
-        .header-section {
-          justify-content: start;
-          align-items: flex-start;
-          flex-direction: column;
-        }
-        .title > div > span:first-of-type {
-          display: block;
-        }
-      }
-    `}</style>
-  </div>
-);
+export { Title, Tabs, LateralMenu };
 
 export default Title;
