+++ frontend/vite.config.js
@@ -26,7 +26,7 @@ module.exports = {
     emptyOutDir: false,
     manifest: true,
     rollupOptions: {
-      input: 'frontend/js/production.js',
+      input: 'js/production.js',
     },
   },
   server: {
