+++ layouts/page-context.ts
@@ -1,0 +1,9 @@
+import React from 'react';
+
+// export const PageContext = React.createContext({ dirigeant: null });
+export const PageContext = React.createContext({
+  dirigeant: {
+    name: 'Dubois',
+    firstName: 'Angela',
+  },
+});
